+++ helper/tool/cluster_client.go
@@ -8,17 +8,12 @@ import (
 
 type (
 	ClusterClient struct {
-		client     *redis.ClusterClient
-		prefix     string
-		baseClient *BaseClient
-		nodeId     string
+		client *redis.ClusterClient
+		prefix string
+		nodeId string
+		IClient
 	}
 )
-
-func (t *ClusterClient) Keys(ctx context.Context, match string) ([]string, error) {
-
-	return t.baseClient.Keys(ctx, match)
-}
 
 func (t *ClusterClient) NodeId(ctx context.Context) string {
 	return t.nodeId
@@ -72,78 +67,4 @@ func (t *ClusterClient) Nodes(ctx context.Context) []Node {
 		result = append(result, nd)
 	}
 	return result
-}
-
-func (t *ClusterClient) KeySpace(ctx context.Context) ([]map[string]any, error) {
-
-	return t.baseClient.KeySpace(ctx)
-
-}
-
-func (t *ClusterClient) Memory(ctx context.Context) (map[string]any, error) {
-
-	return t.baseClient.Memory(ctx)
-
-}
-
-func (t *ClusterClient) CommandStats(ctx context.Context) ([]map[string]any, error) {
-
-	return t.baseClient.CommandStats(ctx)
-}
-
-func (t *ClusterClient) Persistence(ctx context.Context) (map[string]any, error) {
-
-	return t.baseClient.Persistence(ctx)
-
-}
-
-func (t *ClusterClient) Server(ctx context.Context) (map[string]any, error) {
-
-	return t.baseClient.Server(ctx)
-
-}
-
-func (t *ClusterClient) Clients(ctx context.Context) (map[string]any, error) {
-
-	return t.baseClient.Clients(ctx)
-
-}
-
-func (t *ClusterClient) Stats(ctx context.Context) (map[string]any, error) {
-
-	return t.baseClient.Stats(ctx)
-
-}
-
-func (t *ClusterClient) DbSize(ctx context.Context) (int64, error) {
-
-	return t.baseClient.DbSize(ctx)
-}
-
-func (t *ClusterClient) Info(ctx context.Context) (map[string]string, error) {
-
-	return t.baseClient.Info(ctx)
-
-}
-
-func (t *ClusterClient) ClientList(ctx context.Context) ([]map[string]any, error) {
-
-	return t.baseClient.ClientList(ctx)
-
-}
-
-func (t *ClusterClient) Object(ctx context.Context, queueName string) (objstr *ObjectStruct, err error) {
-
-	return t.baseClient.Object(ctx, queueName)
-
-}
-
-func (t *ClusterClient) ZCard(ctx context.Context, key string) (int64, error) {
-
-	return t.baseClient.ZCard(ctx, key)
-
-}
-
-func (t *ClusterClient) Monitor(ctx context.Context) (string, error) {
-	return t.baseClient.Monitor(ctx)
 }
